# frontend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Deploy Hub API
  version: 0.0.1
paths:
...
  /api/v1/partners/:
    get:
      operationId: v1_partners_list
      description: GET is the operator list (no secrets). POST is T1 partner.create.
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerList'
          description: ''
    post:
      operationId: v1_partners_create
      description: GET is the operator list (no secrets). POST is T1 partner.create.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PartnerCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PartnerCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerCreateResult'
          description: ''
  /api/v1/partners/{id}/:
    get:
      operationId: v1_partners_retrieve
      description: GET never echoes hubk_ or whsec_.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerPublic'
          description: ''
...
components:
  schemas:
...
    CandidateTarget:
      type: object
      properties:
        id:
          type: integer
        host:
          type: string
        kind:
          type: string
        tunnel:
          type: boolean
      required:
      - host
      - id
      - kind
      - tunnel
...
    IntakeStatus:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/StatusEnum'
        mode:
          $ref: '#/components/schemas/IntakeStatusModeEnum'
        configured:
          type: boolean
        as_of:
          type: string
          nullable: true
      required:
      - configured
      - mode
      - status
    IntakeStatusModeEnum:
      enum:
      - fake
      - configured
      type: string
      description: |-
        * `fake` - fake
        * `configured` - configured
...
    PartnerCreate:
      type: object
      properties:
        slug:
          type: string
          maxLength: 64
          pattern: ^[-a-zA-Z0-9_]+$
        name:
          type: string
          default: ''
          maxLength: 128
        confirm_name:
          type: string
      required:
      - confirm_name
      - slug
    PartnerCreateResult:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        hubk:
          type: string
        whsec:
          type: string
      required:
      - hubk
      - id
      - name
      - slug
      - whsec
    PartnerDestination:
      type: object
      properties:
        id:
          type: integer
        host:
          type: string
        kind:
          type: string
      required:
      - host
      - id
      - kind
    PartnerList:
      type: object
      properties:
        partners:
          type: array
          items:
            $ref: '#/components/schemas/PartnerPublic'
        intake:
          $ref: '#/components/schemas/IntakeStatus'
        api_enabled:
          type: boolean
        candidate_targets:
          type: array
          items:
            $ref: '#/components/schemas/CandidateTarget'
      required:
      - api_enabled
      - candidate_targets
      - intake
      - partners
    PartnerPublic:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        destination_order:
          type: array
          items:
            type: integer
        destinations:
          type: array
          items:
            $ref: '#/components/schemas/PartnerDestination'
        suspended:
          type: boolean
        site_ids:
          type: array
          items:
            type: integer
      required:
      - destination_order
      - destinations
      - id
      - name
      - site_ids
      - slug
      - suspended
...
    StatusEnum:
      enum:
      - degraded
      - error
      type: string
      description: |-
        * `degraded` - degraded
        * `error` - error
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid